% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{CalculateCumulativeBurden}
\alias{CalculateCumulativeBurden}
\title{Calculate the cumulative burden}
\usage{
CalculateCumulativeBurden(Burdens)
}
\arguments{
\item{Burdens}{A list containing vectors of burdens for all units from each previous survey}
}
\description{
Calculates the cumulative burden for all units
}
\examples{
RunSimulation (FrameNames = rep(c("SampleFrame1","SampleFrame2"), 5),
                         DesignNames = rep(c("SampleDesign1", "SampleDesign2"), 5),
                         DatabaseNames = rep(c("SampleDatabase1", "SampleDatabase2"), 5),
                         BurdenFactors = rep(c(1, 5, 4, 3, 7), 2),
                         CoordinationMethods = c("Microstrata", "GlobalNegative", "CottonHesse", "None"))
lapply(SurveyData.Get (CumulativeBurden), CalculateCumulativeBurden)
}

