% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{CompareBurdens}
\alias{CompareBurdens}
\title{Compare cumulative burdens}
\usage{
CompareBurdens(CumulativeBurden)
}
\arguments{
\item{CumulativeBurden}{A list of cumulative burdens from the SurveyData environment}
}
\description{
Compare the cumulative burdens of coordination methods produced from a simulation
}
\examples{
RunSimulation (FrameNames = rep(c("SampleFrame1","SampleFrame2"), 5),
                         DesignNames = rep(c("SampleDesign1", "SampleDesign2"), 5),
                         DatabaseNames = rep(c("SampleDatabase1", "SampleDatabase2"), 5),
                         BurdenFactors = rep(c(1, 5, 4, 3, 7), 2),
                         CoordinationMethods = c("Microstrata", "GlobalNegative", "CottonHesse", "None"))
CompareBurdens (SurveyData.Get (CumulativeBurden))
}

