% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{CreateSurveyFrame}
\alias{CreateSurveyFrame}
\title{Create a survey data frame}
\usage{
CreateSurveyFrame(SurveyNumber)
}
\arguments{
\item{SurveyNumber}{The survey which information should be retrieved from}
}
\description{
Builds a survey data frame containing the StrataID, RandomNumbers, Burdens 
             and UnitId vectors in the SurveyData environment
}
\keyword{internal}

