% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{GetSurveySequence}
\alias{GetSurveySequence}
\title{Gets the current survey sequence}
\usage{
GetSurveySequence(SurveyNumber, CoordinationMethod)
}
\arguments{
\item{SurveyNumber}{The current survey which is being run}

\item{Coordination}{method Either "CottonHesse", "GlobalNegative" or "Microstrata".
If something else is provided, such as "NoCoordination", no coordination is used}
}
\description{
The survey sequence will be different depending on the coordination CoordinationMethod
}
\examples{
GetSurveySequence (5, "GlobalNegative")
}

