% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{RunSimulation}
\alias{RunSimulation}
\title{Run a sampling simulation}
\usage{
RunSimulation(FrameNames, DesignNames, DatabaseNames, BurdenFactors,
  CoordinationMethods)
}
\arguments{
\item{FrameNames}{The names of the R data frames containing the survey frames}

\item{DesignNames}{The names of the R data frames containing the survey designs}

\item{DatabaseNames}{The names of the R data frames containing the survey databases}

\item{BurdenFactors}{The burden associated with each survey}

\item{CoordinationMethods}{The coordination methods to use (valid options are "GlobalNegative", "Microstrata" and "CottonHesse")}
}
\description{
Runs a sampling simulation and produces the cumulative burden
             for each of the specified coordination methods
}
\details{
Input: FrameNames, DesignNames and DatabaseNames are objects in SurveyData environment;
         Ouput: CumulativeBurden in SurveyData environment
}
\examples{
RunSimulation (FrameNames = rep(c("SampleFrame1","SampleFrame2"), 5),
                         DesignNames = rep(c("SampleDesign1", "SampleDesign2"), 5),
                         DatabaseNames = rep(c("SampleDatabase1", "SampleDatabase2"), 5),
                         BurdenFactors = rep(c(1, 5, 4, 3, 7), 2),
                         CoordinationMethods = c("Microstrata", "GlobalNegative", "CottonHesse", "None"))
}

