% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{RunSurvey}
\alias{RunSurvey}
\title{Run a survey with sample coordination}
\usage{
RunSurvey(FrameName = "SampleFrame1", DesignName = "SampleDesign1",
  DatabaseName = "SampleDatabase1", SurveyNumber = 1, BurdenFactor = 1,
  CoordinationMethod = "CottonHesse")
}
\arguments{
\item{FrameName}{The name of the R data frame containing the survey frame}

\item{DesignName}{The name of the R data frame containing the survey design}

\item{DatabaseName}{The name of the R data frame containing the survey database}

\item{SurveyNumber}{The current survey number}

\item{BurdenFactor}{The burden associated with the current survey}

\item{CoordinationMethod}{The coordination method to use (valid options are "GlobalNegative" and "CottonHesse")}
}
\description{
Coordinates a survey with previously RunSurvey surveys and selects units for the sample
}
\details{
Input: FrameName, DesignName and DatabaseName objects in SurveyData environment;
         Ouput: Strata1d, UnitID, RandomNumbers, Burdens, StrataSizes, NumberUnitsSelected,
                Burdens in SurveyData environment
}
\examples{
RunSurvey ()
SurveyData.Get (Burdens)
}

