% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleDatabase1}
\alias{SampleDatabase1}
\title{[DATA] Sample Database 1}
\format{An object of class \code{data.frame} with 100 rows and 2 columns.}
\usage{
SampleDatabase1
}
\description{
Sample Database 1. Contains two columns: UnitID, RandomNumbers.
}
\keyword{datasets}

