% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleDatabase2}
\alias{SampleDatabase2}
\title{[DATA] Sample Database 2}
\format{An object of class \code{data.frame} with 100 rows and 2 columns.}
\usage{
SampleDatabase2
}
\description{
Sample Database 2. Contains two columns: UnitID, RandomNumbers.
}
\keyword{datasets}

