% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleDesign1}
\alias{SampleDesign1}
\title{[DATA] Sample Design 1}
\format{An object of class \code{data.frame} with 2 rows and 2 columns.}
\usage{
SampleDesign1
}
\description{
Sample Design 1. Contains two columns: StratID, NumberUnitsSelected.
}
\keyword{datasets}

