% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleDesign2}
\alias{SampleDesign2}
\title{[DATA] Sample Design 2}
\format{An object of class \code{data.frame} with 3 rows and 2 columns.}
\usage{
SampleDesign2
}
\description{
Sample Design 2. Contains two columns: StratID, NumberUnitsSelected.
}
\keyword{datasets}

