% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleFrame1}
\alias{SampleFrame1}
\title{[DATA] Sample Frame 1}
\format{An object of class \code{data.frame} with 100 rows and 2 columns.}
\usage{
SampleFrame1
}
\description{
Sample Frame 1. Contains two columns: UnitID, StratID.
}
\keyword{datasets}

