% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleFrame2}
\alias{SampleFrame2}
\title{[DATA] Sample Frame 2}
\format{An object of class \code{data.frame} with 100 rows and 2 columns.}
\usage{
SampleFrame2
}
\description{
Sample Frame 2. Contains two columns: UnitID, StratID.
}
\keyword{datasets}

