% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{SurveyData.Assign}
\alias{SurveyData.Assign}
\title{Assign to SurveyData environment}
\usage{
SurveyData.Assign(ObjectName, ObjectValue = NA, DeparseObjectName = FALSE)
}
\arguments{
\item{ObjectName}{The variable name to assign}

\item{ObjectValue}{The value to assign to the variable}

\item{DeparseObjectName}{true if ObjectName is a string and deparse needs to be called, false otherwise}
}
\description{
Assigns or overrides an existing variable in the SurveyData environment
}
\details{
By default, the value NA is assigned when unspecified
}
\keyword{internal}

