% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{SurveyData.Get}
\alias{SurveyData.Get}
\title{Get from SurveyData environment}
\usage{
SurveyData.Get(ObjectName, DeparseObjectName = FALSE)
}
\arguments{
\item{ObjectName}{The variable name to get the value of}

\item{DeparseObjectName}{true if ObjectName is a string and deparse needs to be called, false otherwise}
}
\description{
Gets the value of a variable defined in the SurveyData environment
}
\details{
The ObjectName argument should be an object name, not a string
}
\keyword{internal}

