% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\docType{data}
\name{SurveyData}
\alias{SurveyData}
\title{Survey data environment}
\format{An object of class \code{environment} of length 6.}
\usage{
SurveyData
}
\description{
Project variables are stored are retrieved from this environment, such as the project directory
}
\details{
To see variables assigned by default use ls(envir=SurveyData)
}
\keyword{internal}

