% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{UpdateSurveyLists}
\alias{UpdateSurveyLists}
\title{Update the survey lists}
\usage{
UpdateSurveyLists(Frame, SurveyNumber = 1, UpdateRandomNumbers = TRUE)
}
\arguments{
\item{Frame}{The frame which contains the RandomNumbers and Burdens vectors}

\item{SurveyNumber}{The survey for which the RandomNumbers and Burdens lists in the 
SurveyData environment should be updated (usually equal to the current survey number)}

\item{UpdateRandomNumbers}{true if RandomNumbers should be updated, false otherwise}
}
\description{
Update the RandomNumbers and Burdens in the SurveyData environment
}
\keyword{internal}

